+++ lib/marv/cli.rb
@@ -5,10 +5,6 @@ module Marv
   module CLI
     class Commands < Project
 
-      def self.source_root
-        ::File.expand_path(::File.join(Marv.root, 'layouts'))
-      end
-
       # Configure marv
       desc "config", "Configure Marv projects and servers"
       long_desc "Creates a global config.rb file that can be used to auto-configure your projects and servers"